Changed the package name to not included an underscore character. Implemented a basic BoardView class and a testing Main class. Tiles now render in a resizable board, special tiles are set a specific color and you can easily assign a type to a tile by number.
