--- laddersgame_xtreme/src/main/java/no/ntnu/idatx2003/exam2025/laddersgame_xtreme/Main.java
@@ -1,7 +1,0 @@
-package no.ntnu.idatx2003.exam2025.laddersgame_xtreme;
-
-public class Main {
-  public static void main(String[] args) {
-    System.out.println("Hello world!");
-  }
-}
