--- laddersgame_xtreme/src/main/java/no/ntnu/idatx2003/exam2025/laddersgame_xtreme/controller/readme.txt
@@ -1,2 +1,0 @@
-Controller folder is for "Java based controller/UI logic"
--> JavaFX backend I assume.
