--- laddersgame_xtreme/src/main/java/no/ntnu/idatx2003/exam2025/laddersgame_xtreme/dao/readme.txt
@@ -1,5 +1,0 @@
-DAO files for interacting with a database.
-
-
-Im going to create player objects that a user can create and then wins, losses and winrate etc can be stored there
-as a database implementation for extra points.
