--- laddersgame_xtreme/src/main/java/no/ntnu/idatx2003/exam2025/laddersgame_xtreme/view/BoardView.java
@@ -1,5 +1,0 @@
-package no.ntnu.idatx2003.exam2025.laddersgame_xtreme.view;
-
-public class BoardView {
-
-}
