+++ laddersgamextreme/src/main/java/no/ntnu/idatx2003/exam2025/laddersgamextreme/model/DennisNotes.txt
